Need to finish add question

+++ app/Http/Controllers/AnswerController.php
@@ -52,7 +52,7 @@ class AnswerController extends Controller
     {
       $answer = Answer::find($id);
 
-      //$this->authorize('delete', $answer);
+      $this->authorize('delete', $answer);
       $answer->delete();
 
       return $answer;
